InsertGame fix

+++ main.py
@@ -171,8 +171,11 @@ def main():
                     Database.InsertGame(currentAirport, energy, maxEnergy, birdName, int(score), gameId=currentGameId)
                     return
                 elif saveAction == "n":
-                    Database.DeleteGame(currentGameId)
-                    return
+                    if(savedGame):
+                        return
+                    else:
+                        Database.DeleteGame(currentGameId)
+                        return
         else:
             print("Wrong input")
